Create ca_revised.py
This copy of the CA code has a revised edges feature so that the established edges are not erased when the first row is initialized. 

This version of the code also creates the CA array with two extra columns so that the edges don't need to be part of the final array, which is returned without column 0 or the last column.